New Room Clearing
Added portals that spawn when the room is cleared, no longer automatically swaps to new room. (This is part of a trial for the sprint, the setting is interchangeable with 2 other room cleared options)
 - Option 1 is a portal that spawns at a random position on the map, when entered the next room loads.
 - Option 2 is multiple portals that spawn in set positions on the map. It has icons showing a reward, so the player chooses which one they want as they go to the next room.

Added a spawn animation to the upgrade scenes.
Added coins moving going to the player on room clear.
Adjusted item rarities.
Reworked the player to spawn in set positions instead of randomly across the map (To prevent getting stuck or taking damage on spawn)