Github & Animation Fix
I accidentally merged the last github commit with an older commit which broke the project. This commit should fix this.

Changed the sprite flipping to be based on which direction they are moving, not their next path position.
Changes to the bomber enemy's stats.
Fixed the bomber enemy's animation not slowing down when stopping to shoot bombs.